AudioService: pause media playback when registered policy client dies
Broadcast becoming noisy intent when the client of a registered dynamic
audio policy dies with an active media playback rerouted from a remote
submix device.
This happens for instance when Android Auto app dies with an active
projection session.
This makes the user experience similar to what happens when disconnecting
the Bluetooth A2DP connection from a car kit.

Bug: 186581483
Test: manual test with head unit simulator
Change-Id: I2d98894145ff87e4711897fd09797c8e941fd0f1