Bypass HUN snooze for notifs with FSI
If an app posts a time-sensitive (alarm, call, etc.) notification that
can either HUN or FSI, it's possible for *neither* alerting behavior to
happen if the user has snoozed HUNs for the package and System UI has
suppressed FSIs because the screen is on and unlocked and the
FSI_REQUIRES_KEYGUARD flag is set.

In this case, assume any notification with an FSI is time-sensitive
(like an alarm or incoming call) and ignore whether HUNs have been
snoozed for the package.

Bug: 245356701
Test: atest NotificationInterruptStateProviderImplTest
Test: manual tests of relevant FSI/HUN snooze scenarios
Change-Id: Idee76ef9fc9d303903617ffd5bc79aad0756b2a0