Introduce ComposeInitializer (1/2)
This CL introduces ComposeInitializer, which can be used by Window root
views to make Compose usable in that window. Once a Window is
initialized, we can simply add ComposeViews to the Window hierarchy and
they will work as expected.

This CL also uses ComposeInitializer in the NotificationShadeWindowView.
This will be used by the Quick Settings footer actions, that have
already been implemented using Compose.

Bug: 230740991
Test: ComposeInitializerTest
Change-Id: I232bd12d3a46083101663a5bc557472263c27c38