RESTRICT AUTOMERGE Center align the lock screen wallpaper
CP for UDC initial release where lockscreen live wallpaper is not enabled, and later we need to revert this for udc-qpr-dev.

On smaller display of a multi display device the lock screen wallpaper
should be showing the center area of the wallpaper. Check the display id
to find out which display the lock screen wallpaper is on, the check
needs to happened when we are adjusting the bounds to get the new
display id.

Bug: 263932616
Bug: 289079608
Test: set the lock screen wallpaper and observe the result
Change-Id: I32b93084a9e716a06e032f83a8cf625c25d3e5e5
(cherry picked from commit a5fc61f77a5715a19f4be67517fac5a25cff3d15)