Prevent removing the target user during a user switch.
This CL prevents removing the target user and allows removing the
user that has been switched away. Let's say we're switching from
userA to userB, before this CL during the switch it was possible to
remove userB and it wasn't possible to remove userA. After this CL
this problem is fixed.

This CL also disables REMOVE_GUEST_ON_EXIT while running
UserLifecycleStressTest.switchToExistingGuestAndStartOverStressTest,
and restores it back after the test.

Bug: 264667155
Test: atest FrameworksServicesTests:com.android.server.pm.UserLifecycleStressTest#switchToExistingGuestAndStartOverStressTest
Change-Id: I802191492adc7a60c7c43ea7651a124d2a8425da