Revert "Enable tethering namespace accessible in Android native level."
This reverts commit 8713a0be58d79d978c8adc189f0bd315a13ee22c.

Reason for revert: Original commit was for native experiment setup.
But SEPolicy was not updated and sync to native level always fail in T devices.
This CL will clean up incomplete native experiment setup.
Even if we decide that we want to use flags from native code in the Tethering module,
we will do so using a new namespace, tethering_u_plus_native.

Test: m
Bug: 264947218
Change-Id: If32fe0b237dcb213886efb72430ea7575ab64530