[ImeStateComputer]: Preserve last visible state for STATE_UNCHANGD
In ImeVisibilityStateComputer#computeState, we only keep the legacy
behavior to do nothing for handling SOFT_INPUT_STATE_CHANGED flag of the
focusing window without changing the requested visiblity.

However, with per-windowState based IME visiblity model, we should
perserve the requested IME visiblity from the last IME targeted window
to the current focusing window in case we are failing the tests related
to launching a new activity with STATE_UNCHANGED from the ativity with
IME shown.

Fix: 265439883
Bug: 246309664
Test: atest InputMethodStressTest:com.android.inputmethod.stresstest\
         .AutoShowTest#autoShow_forwardWithKeyboardOn
    Expect it can passed on the following test parameters combination
      - windowFocusFlags=0 /* no windowFocusFlags */
      - softInputVisibility=1 /* STATE_UNCHANGED flag */
      - softInputAdjustment=any() /* SOFT_INPUT_ADJUST_* flags */

Change-Id: I74907f27f275ba6028b4683145205f2cce2b2f45