Remove not necessary constructor of CallEndpointException
Some constructor in CallEndpointException are removed because the server side has no scenario to use.
CallEndpointException was introduced together with b/262452554 and it is used to report exceptions that can be occurred during change of CallEndpoint. But some constructor are not necessary because we can create it like CallEndpointController#getErrorResult.

bug: 264468307
Test: build and manual test
Change-Id: I24bc13b17d97c79f6c3b8fdce46ea7ac55c90f21