Simplify initialization of transition controller
RootWindowContainer#mTransitionController was always null because
setWindowManager(create TransitionController) is called after the
creation of RootWindowContainer.

Now TransitionController is created with ActivityTaskMangerService,
and its fields of window manager dependencies are assigned from
setWindowManager. Then all WindowContainer#mTransitionController
will be non-null.

This also fixes that setNavigationBarLumaSamplingEnabled was
missing for transient launch because mStatusBar is always null
by the initialization order.

Bug: 163976519
Test: atest TransitionTests

Change-Id: Ic72c3f70b6de005d9b68a43e53510c2a5f4bdb44