Clarify start/stop DTMF tone API documentation.
The API docs are not clear; they insinuate that a caller can issue any
number of "play" DTMF requests followed by a single "stop" request and
that the framework will interpret this as two digits in succession.
In reality, the documentation is referring to the fact that the local tone
will correctly stop if another tone is already playing.  The API
requests are sent directly to the modem as the InCallService sends them,
and it is possible that some ConnectionService implementations (more
specifically some vendor IMS implementations) will get confused if
they receiver multiple starts and then a single stop.

Test: Docs only change
Bug: 262699364
Change-Id: I770008cc3b000f739d930ea899df5e98d1b17129